Rewrite for easier extensions.
After the swift-language argparser I made (in another repo on github)
I decided to update the java implementation to match it.  Now it should
be easier to define exotic parameter types.

I still need to implement the help text formatting, and update the
javadoc extensively.  Then it's going on maven central.