Take Iterable<String> instead of Collection<String>, Parser can skip args.
It's less restrictive to take Iterable<> rather than Collection<>, and
I realized that in many places an Iterable<> was all I needed.

Also, for the Parser, I realized that many programs strip a subcommand
off the args array prior to parsing.  This means constructing a copied
array, which is wasteful. So now, the Parser also takes a `skip` integer
argument to start at an offset into the argument array.  No more copying needed.

These are incompatible changes, though, so I need to bump the next version
number to 2.0.0