Further enhancements for org.rwtodd.args 2.0.0
Now Param can be one of three types... No-Arg, One-Arg, and Decorative.

Decorative Params merely live in the param list, waiting to make the
help text output nicer.  They are not allowed in the Parer's lookup map.
Right now the only supplied DecorativeParam is the ParamGroup, which
puts a heading above a set of parameters.

This commit also adds an AppendingParam, which takes any One-Arg param
and allows it to appear multiple times in the command-line. Each argument
will be appended to a list.

AccumulatingParam has been renamed to CountingParam, since that's really
what it does.